style: Group classes per module. Replace @autowired with DI
