--- backend/src/main/java/com/example/demoio/controllers/Home.java
@@ -1,19 +1,0 @@
-package com.example.demoio.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/home")
-public class Home extends BaseController {
-
-    @GetMapping()
-    public String showHomepage(Model model) {
-
-
-
-        return "home";
-    }
-}
