+++ backend/src/main/java/com/example/demoio/core/auth/services/CustomUserDetails.java
@@ -1,0 +1,29 @@
+package com.example.demoio.core.auth.services;
+
+import com.example.demoio.models.User;
+import com.example.demoio.modules.datastorage.repositories.UserRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+
+@Service
+public class CustomUserDetails implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    public CustomUserDetails(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = userRepository.findByUsername(username);
+        if (user == null) {
+            throw new UsernameNotFoundException("User not found");
+        }
+        return new org.springframework.security.core.userdetails.User(user.getUsername(), user.getPassword(), new ArrayList<>());
+    }
+}
