--- backend/src/main/java/com/example/demoio/models/dto/UpdateRanking.java
@@ -1,4 +1,0 @@
-package com.example.demoio.models.dto;
-
-public record UpdateRanking(int gameID, int score) {
-}
