--- backend/src/main/java/com/example/demoio/models/dto/UpdateUserCoinsRequest.java
@@ -1,4 +1,0 @@
-package com.example.demoio.models.dto;
-
-public record UpdateUserCoinsRequest(int coins) {
-}
