--- backend/src/main/java/com/example/demoio/models/dto/UserDataResponse.java
@@ -1,4 +1,0 @@
-package com.example.demoio.models.dto;
-
-public record UserDataResponse(String username, double userCoins) {
-}
