--- backend/src/main/java/com/example/demoio/models/orm/GameProgress.java
@@ -1,29 +1,0 @@
-package com.example.demoio.models.orm;
-
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-public class GameProgress {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private int game_id;
-    private int bestScore;
-
-    @ManyToOne()
-    @JoinColumn(name = "user_fk")
-    private User user;
-
-    public GameProgress(int game_id, int bestScore) {
-        this.game_id = game_id;
-        this.bestScore = bestScore;
-    }
-}
