--- backend/src/main/java/com/example/demoio/models/orm/User.java
@@ -1,34 +1,0 @@
-package com.example.demoio.models.orm;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private int id;
-
-    private String username;
-    private String password;
-    private int userCoins;
-    private int totalUserScore;
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
-    private List<GameProgress> gameProgressList = new ArrayList<>();
-
-    public User(String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
-}
