--- backend/src/main/java/com/example/demoio/models/thymeleaf/DailyTask.java
@@ -1,4 +1,0 @@
-package com.example.demoio.models.thymeleaf;
-
-public record DailyTask(String gameName, String imageSlugName, String taskDescription, String callbackValue) {
-}
