--- backend/src/main/java/com/example/demoio/models/thymeleaf/DisplayRankingData.java
@@ -1,4 +1,0 @@
-package com.example.demoio.models.thymeleaf;
-
-public record DisplayRankingData(String username, int score) {
-}
