+++ backend/src/main/java/com/example/demoio/modules/app/dto/UserDataResponse.java
@@ -1,0 +1,4 @@
+package com.example.demoio.modules.app.dto;
+
+public record UserDataResponse(String username, double userCoins) {
+}
