+++ backend/src/main/java/com/example/demoio/modules/dailytasks/dto/DailyTask.java
@@ -1,0 +1,4 @@
+package com.example.demoio.modules.dailytasks.dto;
+
+public record DailyTask(String gameName, String imageSlugName, String taskDescription, String callbackValue) {
+}
