+++ backend/src/main/java/com/example/demoio/modules/dailytasks/dto/DisplayRankingData.java
@@ -1,0 +1,4 @@
+package com.example.demoio.modules.dailytasks.dto;
+
+public record DisplayRankingData(String username, int score) {
+}
