+++ backend/src/main/java/com/example/demoio/modules/datastorage/repositories/UserRepository.java
@@ -1,0 +1,15 @@
+package com.example.demoio.modules.datastorage.repositories;
+
+import com.example.demoio.models.User;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+
+public interface UserRepository extends CrudRepository<User, Long> {
+
+    User findByUsername(String username);
+
+    @Query("SELECT u FROM User u ORDER BY u.totalUserScore DESC LIMIT 10")
+    List<User> findTop10ByScore();
+}
