+++ backend/src/main/java/com/example/demoio/modules/ranking/dto/UpdateRanking.java
@@ -1,0 +1,4 @@
+package com.example.demoio.modules.ranking.dto;
+
+public record UpdateRanking(int gameID, int score) {
+}
