+++ backend/src/main/resources/templates/daily-tasks.html
@@ -20,7 +20,7 @@
 						>.
 					</p>
 
-                    <!--/*@thymesVar id="task" type="com.example.demoio.models.thymeleaf.DailyTask"*/-->
+					<!--/*@thymesVar id="task" type="com.example.demoio.dailytasks.dto.DailyTask"*/-->
 					<div
 						th:each="task : ${availableTasks}"
 						class="mx-auto mb-6 w-[600px] rounded-xl bg-neutral-100 p-5"
