--- backend/src/test/java/com/example/demoio/DemoIoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.demoio;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoIoApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
